Move all pytest code to nltk/test
