+++ nltk/test/metrics.doctest
@@ -5,6 +5,13 @@
 Metrics
 =======
 
+------------------------
+Check Numpy Is Installed
+------------------------
+    >>> import pytest
+    >>> _ = pytest.importorskip("numpy")
+
+
 The `nltk.metrics` package provides a variety of *evaluation measures*
 which can be used for a wide variety of NLP tasks.
